requirements.txt
absl-py=2.1.0=pypi_0
ag2=0.6.0=pypi_0
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.11.6=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
anthropic=0.40.0=pypi_0
anyio=4.6.2.post1=pypi_0
asgiref=3.8.1=pypi_0
asttokens=3.0.0=pypi_0
astunparse=1.6.3=pypi_0
asyncer=0.0.8=pypi_0
asyncio-atexit=1.0.1=pypi_0
attrs=24.2.0=pypi_0
autogen=0.6.0=pypi_0
autogen-agentchat=0.4.0.dev6=pypi_0
autogen-core=0.4.0.dev6=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.2.1=pypi_0
beautifulsoup4=4.12.3=pypi_0
blinker=1.9.0=pypi_0
build=1.2.2.post1=pypi_0
bzip2=1.0.8=h2466b09_7
ca-certificates=2024.8.30=h56e8100_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
chroma-hnswlib=0.7.6=pypi_0
chromadb=0.5.23=pypi_0
click=8.1.7=pypi_0
codebleu=0.7.0=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pypi_0
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
dash=2.18.2=pypi_0
dash-bootstrap-components=1.6.0=pypi_0
dash-core-components=2.0.0=pypi_0
dash-daq=0.5.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
dash-uploader=0.6.1=pypi_0
dataclasses-json=0.6.7=pypi_0
datasets=3.2.0=pypi_0
debugpy=1.8.11=pypi_0
decimer=2.7.1=pypi_0
decimer-segmentation=1.4.0=pypi_0
decorator=5.1.1=pypi_0
deprecated=1.2.15=pypi_0
dill=0.3.8=pypi_0
dirtyjson=1.0.8=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
docker=7.1.0=pypi_0
durationpy=0.9=pypi_0
efficientnet=1.1.1=pypi_0
evaluate=0.4.3=pypi_0
executing=2.1.0=pypi_0
fastapi=0.115.6=pypi_0
filelock=3.16.1=pypi_0
filetype=1.2.0=pypi_0
flaml=2.3.2=pypi_0
flask=3.0.3=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.55.0=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.9.0=pypi_0
gast=0.4.0=pypi_0
google-auth=2.37.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
googleapis-common-protos=1.66.0=pypi_0
greenlet=3.1.1=pypi_0
grpcio=1.69.0=pypi_0
grpcio-tools=1.69.0=pypi_0
h11=0.14.0=pypi_0
h2=4.1.0=pypi_0
h5py=3.12.1=pypi_0
hpack=4.0.0=pypi_0
httpcore=1.0.7=pypi_0
httptools=0.6.4=pypi_0
httpx=0.27.2=pypi_0
httpx-sse=0.4.0=pypi_0
huggingface-hub=0.27.0=pypi_0
humanfriendly=10.0=pypi_0
hyperframe=6.0.1=pypi_0
idna=3.10=pypi_0
imageio=2.36.0=pypi_0
importlib-metadata=8.4.0=pypi_0
importlib-resources=6.4.5=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.30.0=pypi_0
itsdangerous=2.2.0=pypi_0
jax=0.4.30=pypi_0
jaxlib=0.4.30=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.4=pypi_0
jiter=0.7.1=pypi_0
joblib=1.4.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
keras=2.12.0=pypi_0
keras-applications=1.0.8=pypi_0
kiwisolver=1.4.7=pypi_0
kubernetes=31.0.0=pypi_0
langchain=0.3.7=pypi_0
langchain-community=0.3.7=pypi_0
langchain-core=0.3.19=pypi_0
langchain-openai=0.2.9=pypi_0
langchain-text-splitters=0.3.2=pypi_0
langsmith=0.1.143=pypi_0
lazy-loader=0.4=pypi_0
libclang=18.1.1=pypi_0
libexpat=2.6.4=he0c23c2_0
libffi=3.4.2=h8ffe710_5
libsqlite=3.47.0=h2466b09_1
libzlib=1.3.1=h2466b09_2
llama-cloud=0.1.6=pypi_0
llama-index=0.12.6=pypi_0
llama-index-agent-openai=0.4.0=pypi_0
llama-index-cli=0.4.0=pypi_0
llama-index-core=0.12.6=pypi_0
llama-index-embeddings-openai=0.3.1=pypi_0
llama-index-indices-managed-llama-cloud=0.6.3=pypi_0
llama-index-llms-openai=0.3.10=pypi_0
llama-index-multi-modal-llms-openai=0.4.0=pypi_0
llama-index-program-openai=0.3.1=pypi_0
llama-index-question-gen-openai=0.3.0=pypi_0
llama-index-readers-file=0.4.1=pypi_0
llama-index-readers-llama-parse=0.4.0=pypi_0
llama-parse=0.5.17=pypi_0
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markdownify=0.14.1=pypi_0
markupsafe=3.0.2=pypi_0
marshmallow=3.23.1=pypi_0
matplotlib=3.9.2=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mem0ai=0.1.44=pypi_0
ml-dtypes=0.2.0=pypi_0
mmh3=5.0.1=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
mypy-extensions=1.0.0=pypi_0
nest-asyncio=1.6.0=pypi_0
networkx=3.4.2=pypi_0
nltk=3.9.1=pypi_0
numpy=1.23.5=pypi_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.20.1=pypi_0
openai=1.58.1=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.4.0=h2466b09_0
opentelemetry-api=1.29.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.29.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.29.0=pypi_0
opentelemetry-instrumentation=0.50b0=pypi_0
opentelemetry-instrumentation-asgi=0.50b0=pypi_0
opentelemetry-instrumentation-fastapi=0.50b0=pypi_0
opentelemetry-proto=1.29.0=pypi_0
opentelemetry-sdk=1.29.0=pypi_0
opentelemetry-semantic-conventions=0.50b0=pypi_0
opentelemetry-util-http=0.50b0=pypi_0
opt-einsum=3.4.0=pypi_0
orjson=3.10.11=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pypi_0
pdf2image=1.17.0=pypi_0
pillow=11.0.0=pypi_0
pillow-heif=0.21.0=pypi_0
pip=24.3.1=pyh8b19718_0
platformdirs=4.3.6=pypi_0
plotly=5.24.1=pypi_0
portalocker=2.10.1=pypi_0
posthog=3.7.4=pypi_0
prompt-toolkit=3.0.48=pypi_0
propcache=0.2.0=pypi_0
protobuf=5.29.3=pypi_0
psutil=6.1.0=pypi_0
pure-eval=0.2.3=pypi_0
pyarrow=19.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pyautogen=0.6.0=pypi_0
pyclipper=1.3.0.post6=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pydantic-settings=2.6.1=pypi_0
pygments=2.18.0=pypi_0
pymupdf=1.24.14=pypi_0
pyparsing=3.2.0=pypi_0
pypdf=5.1.0=pypi_0
pypdf2=3.0.1=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.2.0=pypi_0
pyreadline3=3.5.4=pypi_0
pystow=0.6.1=pypi_0
python=3.11.10=hce54a09_3_cpython
python-bidi=0.6.3=pypi_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2024.2=pypi_0
pywin32=308=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=pypi_0
qdrant-client=1.13.0=pypi_0
rdkit-pypi=2022.9.5=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
retrying=1.3.4=pypi_0
rich=13.9.4=pypi_0
rsa=4.9=pypi_0
scikit-image=0.24.0=pypi_0
scipy=1.14.1=pypi_0
selfies=2.1.2=pypi_0
setuptools=75.5.0=pyhff2d567_0
shapely=2.0.6=pypi_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlalchemy=2.0.35=pypi_0
stack-data=0.6.3=pypi_0
starlette=0.41.3=pypi_0
striprtf=0.0.26=pypi_0
sympy=1.13.1=pypi_0
tenacity=9.0.0=pypi_0
tensorboard=2.12.3=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.12.0=pypi_0
tensorflow-estimator=2.12.0=pypi_0
tensorflow-intel=2.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.5.0=pypi_0
tifffile=2024.9.20=pypi_0
tiktoken=0.8.0=pypi_0
tk=8.6.13=h5226925_1
tokenizers=0.20.3=pypi_0
torch=2.0.1=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.67.0=pypi_0
traitlets=5.14.3=pypi_0
tree-sitter=0.22.3=pypi_0
tree-sitter-python=0.23.1=pypi_0
typer=0.15.1=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.2=pypi_0
ucrt=10.0.22621.0=h57928b3_1
urllib3=2.2.3=pypi_0
uvicorn=0.34.0=pypi_0
vc=14.3=ha32ba9b_23
vc14_runtime=14.42.34433=he29a5d6_23
vs2015_runtime=14.42.34433=hdffcdeb_23
watchfiles=1.0.3=pypi_0
wcwidth=0.2.13=pypi_0
websocket-client=1.8.0=pypi_0
websockets=14.1=pypi_0
werkzeug=3.0.6=pypi_0
wheel=0.45.0=pyhd8ed1ab_0
wrapt=1.14.1=pypi_0
xxhash=3.5.0=pypi_0
xz=5.2.6=h8d14728_0
yarl=1.17.2=pypi_0
zipp=3.21.0=pypi_0
